+++ src/test/kotlin/team_json_delivery/sns_b/domain/post/controller/PostControllerTest.kt
@@ -13,7 +13,7 @@ import team_json_delivery.sns_b.domain.post.model.dto.PostDto
 import team_json_delivery.sns_b.domain.post.service.GetPostService
 
 @WebMvcTest(GetPostController::class)
-class GetPostControllerTest(
+class PostControllerTest(
     private val mockMvc: MockMvc,
     @MockitoBean
     private val getPostService: GetPostService
